[refactor] loading screen component.
