+++ src/Components/ScreenLoadingComponent/ScreenLoadingComponent.tsx
@@ -3,7 +3,7 @@ import { ActivityIndicator } from 'react-native-paper';
 
 import { colors } from '../../styles/colors';
 
-export default function LoadingComponent() {
+export default function ScreenLoadingComponent() {
   return (
     <View style={{ marginTop: 20 }}>
       <ActivityIndicator size="small" color={colors.midRed} />
