+++ src/app/(BottomTabNavigation)/AllCases/CaseScreen/[caseUid].tsx
@@ -1,7 +1,6 @@
 import { useLocalSearchParams } from 'expo-router';
 import React, { useEffect, useState } from 'react';
 import { View, ScrollView, Text } from 'react-native';
-import { ActivityIndicator } from 'react-native-paper';
 
 import styles from './styles';
 import CaseSummaryCard from '../../../../Components/CaseSummaryCard/CaseSummaryCard';
@@ -10,7 +9,7 @@ import ClaimStatusBar from '../../../../Components/ClaimStatusBar/ClaimStatusBar
 import EducationalBar from '../../../../Components/EducationalBar/EducationalBar';
 import EligibilityCard from '../../../../Components/EligibilityCard/EligibilityCard';
 import FormsCard from '../../../../Components/FormsCard/FormsCard';
-import LoadingComponent from '../../../../Components/LoadingComponent/LoadingComponent';
+import ScreenLoadingComponent from '../../../../Components/ScreenLoadingComponent/ScreenLoadingComponent';
 import StatusUpdatesBar from '../../../../Components/StatusUpdatesBar/StatusUpdatesBar';
 import { useCaseContext } from '../../../../context/CaseContext';
 import { fonts } from '../../../../styles/fonts';
@@ -45,8 +44,7 @@ function CaseScreen() {
   return (
     <View style={device.safeArea}>
       {!caseData || !status ? (
-        // <Text>Loading...</Text>
-        <LoadingComponent />
+        <ScreenLoadingComponent />
       ) : (
         <ScrollView
           style={styles.outerScroll}
