+++ src/app/(BottomTabNavigation)/AllCases/FileClaim/[caseUid].tsx
@@ -11,7 +11,7 @@ import {
   ButtonBlack,
   ButtonWhite,
 } from '../../../../Components/AuthButton/AuthButton';
-import LoadingComponent from '../../../../Components/LoadingComponent/LoadingComponent';
+import ScreenLoadingComponent from '../../../../Components/ScreenLoadingComponent/ScreenLoadingComponent';
 import { useCaseContext } from '../../../../context/CaseContext';
 import { fonts } from '../../../../styles/fonts';
 import { device } from '../../../../styles/global';
@@ -58,7 +58,7 @@ export default function FileClaimScreen() {
   return (
     <View style={device.safeArea}>
       {caseData === undefined ? (
-        <LoadingComponent />
+        <ScreenLoadingComponent />
       ) : (
         <View style={instruction.screenContainer}>
           <View style={instruction.contentContainer}>
