+++ src/app/(BottomTabNavigation)/AllCases/OptOut/[caseUid].tsx
@@ -10,7 +10,7 @@ import {
   ButtonBlack,
   ButtonWhite,
 } from '../../../../Components/AuthButton/AuthButton';
-import LoadingComponent from '../../../../Components/LoadingComponent/LoadingComponent';
+import ScreenLoadingComponent from '../../../../Components/ScreenLoadingComponent/ScreenLoadingComponent';
 import { fonts } from '../../../../styles/fonts';
 import { device } from '../../../../styles/global';
 import { instruction } from '../../../../styles/instruction';
@@ -43,7 +43,7 @@ export default function OptOutScreen() {
   return (
     <View style={device.safeArea}>
       {caseData === undefined ? (
-        <LoadingComponent />
+        <ScreenLoadingComponent />
       ) : (
         <View style={instruction.screenContainer}>
           <View style={instruction.contentContainer}>
